Bug fixes
Added detection of the game version to allow for differences between A15
and A16.

Altered /near command so it doesn't send admins under the map.  Will
revert when the Pimps bug is fixed.

Altered recording of zombies for gimme so that it won't spawn the new
behemoth, ferals or radiated zombies.

Small fixes in the chat parser and moved the execution of
gmsg_custom.lua to the top to allow it to override any command.

Fixed bug in waypoint teleport.

Altered key in gimmeZombies table and added new field to help with
removing zombies dropped in A16.

Stopped scanning for noclipping players in A16 until the Pimps fix that
bug. It was causing innocent and very stuck players to get banned.

Fixed issue with deaths still not registering when the starvation mod is
present.  Copied humorous death notices from the pvp section.