Update for A16 B112
The disabling of scanning for players flying under the map is now A16
version specific so it will only be disabled if you have A16 B105.

The gimme zombies are fixed and spawning properly with all the hard new
zombies blocked from being spawned by the bot (for now).

Fixed a minor bug in the chat parser and chat logging.