Bug fixes, new features and misc changes
Changelog

Bugs fixed:

Custom colours for player names wasn't saving and would be lost when the bot restarted.

The IRC command 'near player joe' is working again.  Some recent changes to it weren't perfect which caused it to break.

The ingame command /restart bot, when used with a password the password wasn't triggering the restart but it worked from IRC.

An exploit was fixed, but don't panic.  Nobody found it and only server owners could abuse it.

Fixed misteak in the help for the command that sets player chat colour to a custom colour.

The old command help used <> brackets and when banning someone, if you included the brackets, you would break the admin xml file which is
not ideal.  The bot now strips them out.  You aren't supposed to type those silly!

While not really a bug, the server welcome and other messages shown to players when they join the server has been changed from being timer based
to waiting for head movement from the player which is a sure indication that they have spawned.  This way they will get those messages as soon as
they spawn even if it takes them ages to get there.

Experimental fixes:

There's a rare issue where a bot will go offline after a server restart and fail to reconnect.  I added an extra offline test and added it
to two timers instead of relying on just one.  Assuming this works, the bot should come back after 5 minutes if the issue happens again.

The command '/fix bot' now collects all known valid items and compiles a list.  It uses that list to then remove invalid items from gimme,
the shop and some other lists.  It also re-applies all alter table statements as I found there was at least one bot with issues caused by
somehow skipping one or more of those (probably it got interrupted).  This makes the command take longer but it will fix those things for sure.

New stuff:

On IRC there's a new performance report that you can run by typing run report.  It will report to the private channel every 30 seconds or so.
It counts the number of commands sent to the server by the bot, the number of telnet lines processed and shows the number of players, the mem
output and the command lag time in seconds.  It won't pinpoint where lag is coming from but you get a bit more info than just how long the lag is.
To stop the report type stop report.  It doesn't impact the bot's performance as it is only counting stuff and I may change it to run full time later.

/disable (or enable) claim scan
Every 45 seconds the bot reads the claims of all ingame players. This can be a lot of data and could impact server performance.
If the bot is reporting server lag frequently, you can disable the timed claim scan.
It will still scan when a player leaves the server and can be commanded to do a scan.

/disable (or enable) screamer alert
By default the bot will warn players when screamers are approaching.  You can disable that warning.

/disable (or enable) airdrop alert
By default the bot will inform players when an airdrop occurs near them.  You can disable the message.

/gimme reset time {number} (In minutes. Default is 120)
Reset everyone's gimme counter after (n) minutes.

/reset chat colour
Resets everyone's custom chat colour back to white.  If you have set group colours such as for donors, admins etc, those colours will be used instead.

There's a new IRC command 'view slots' that gives you a report on the current state of reserved slots including a list of all players occupying a slot.

Misc changes:

The functions LookupPlayer and LookupOfflinePlayer now return 0 instead of nil if no player matches the search.