New commands, new daily log, better info.
The command /test as player now accepts a number (seconds).  The default is 300 (5 minutes).  There is also a new command /restore admin.  The bot will restore your admin status immediately.

There is a new daily alerts log which records almost all alerts sent to the IRC alerts channel.

The bot now includes the server date and time and the game day and time in all most log entries to assist with finding them in the telnet logs.

Added a new variation on the hacker alert for flying/clipping players.  The bot will report stationary as well as moving players that are detected flying or clipping.  Also its tracking of player movement while clipping is improved.

The IRC who command now reports what location players are in.