Bugfixes, new command
Fixed a couple of issues with teleports being false flagged as not legit.  This could be caused by the property HideCommandExecutionLog in the server config not being set to 0 or by admins using the console to teleport players rather than using a bot command.

Added new command /claim vote.  It isn't quite working yet but I'll test it and get it working.  It checks for a vote from the player on the website 7daystodie-servers.com and if they have an unclaimed vote there it will spawn a supply crate beside them.