Bugfix
Recent changes that enabled mixed-case location names had an unintended side-effect of breaking the lobby feature if the lobby location was called Lobby.  This was due to Lua being case sensitive and my change not being thorough enough to catch it.

Sorry if this bug stopped your lobby from working.  It should work now.