Bug fixes and new command
You can now set the threshold for drop mining warnings.  This alert isn't perfect in that all it does is count blocks falling off the world and in testing, dropping an entire prefab logged 7 blocks.  If you set this very low it may report stuff that isn't strictly drop mining or bad for the server.  Play with it and see what works best.  You can disable it by setting it to zero.

There was a bug in the bot's automatic relog code that caused the bot to be slow to reconnect after a reboot and it would cause the bot to restart itself.

Part of the lobby fix from earlier managed to not get included thanks to an odd behaviour of my editor.