Bugfix and a little more info on player spawn
I have been battling a little logic bug where the bot is supposed to rejoin the server if it falls off.  Most of the time this works but sometimes it fails to rejoin.  This should fix that bug finally.

Also added a little more info when a player joins the server.  In the IRC channels it now tells you where they spawn and if they get moved to a location such as lobby.