Bug fixes, performance tweaks, new commands
Added command cooldown that applies to players only.  Set it with /set command cooldown 10 (for a 10 second wait before the bot will accept another command per player).

Altered how the bot detects when it goes offline. The bot no longer needs to slowly spam the gt command to test that it is still online.

The bot spams commands much less and much slower when it first starts up.  There was a double up of some commands and it no longer re-sends them every time it relogs.

Added support for the CSMM Patron's Mod.

Added command help for all of the bot's IRC commands.  Many still need content added, but once that is done, there shouldn't be any major collections of undocumented commands left.

The bot now sends the save world command to the server every 10 minutes.