Bug fixes and improvements
There was some code in the reconnect timer that was causing the bot to reconnect to telnet a lot and unnecessarily when the server is idle.  I also tweaked how it keeps tabs on its online state and the state of the server.

In API mode, level 0 server owners now have full access to server console commands from the IRC.  There shouldn't be many console commands that don't get piped back to IRC.

I also added a file check to prevent the API code trying to process empty command outputs.