Bugfixes mainly
Added a new ingame only command /bot info.  It lists some basic info about the bot.

Fixed some issues with the bot's code concerning rebooting the server.  It was sending sa and shutdown extra times due to some code triggering when it wasn't needed.  Fixed it with a serverRebooting flag.  Also the /cancel reboot command now cancels the final countdown as well (though it always stopped the actual reboot).

Added a check for bot updates every time the server reboots so that updates can go out a bit sooner.

Added an extra offline test so that in API mode, if the bot detects players and hasn't seen any telnet traffic in 90 seconds or the bot thinks it is offline, it will reconnect to telnet.

Various tweaks to the match_all function to cope with the different requirements of the API versus telnet since both call this function.

Tweaks to the webAPI code mainly to properly handle piping output to IRC when required.