Bug fixes
I discovered one server had a different carriage return delimiter in Allocs API outputs.  That will be due to that server running on Linux while most including my dev server run on Windows.  I added a function to detect the difference and correctly split lines on the local CRLF characters.

There were 3 lines missing from the inventory scanner that I ported to the API code.

Timeout won't teleport players before the bot thinks/sees that they have spawned in the world.  Note that if the bot is restarted, it flags all ingame players as spawned.  Later I'll refine that so it works that out from the live data rather than making an assumption on bot startup.

Added an IRC command, use telnet.  This allows you to switch the bot back to using telnet if for any reason commands stop working ingame. Due to the CRLF bug and not re-reading admin list after a change to admins, the bot could stop recognising some or all admins as admins even though the server had the correct data.  That was a fun bug to fix xD