Bugfixes and new stuffs
New command /enable (or disable) level hack alert.  On some modded servers player levels increase by very large amounts so the level hack warning is just a nuisance.

New command /enable (or disable) bot command log.  For debugging purposes mainly, you can have the bot log every command it sends to the server.  It filters out or just doesn't log lines with sensitive content such as passwords etc.  All of the bot's logs use the server's current time so you should be able to match up anything in the logs with the telnet log.

New command /location hide (or unhide) {location name}.  This will hide locations from the /locations command.  To stop them displaying a welcome message, make them private (the default for new locations).

New command /location lobby/not lobby {location name}.  You can now make any location act as the lobby.  If you set this on a location, the bot will unset it for all other locations automatically.  Note that if you have a location called lobby, that will always be the lobby and cannot be overridden by doing this.

I disabled a trigger that reacts to the admin list updating.  It is also triggered by the webpermissions updating.  Now that is detected and handled in the match_all function and is aware of the bot's last command so it won't trigger on webtokens updating.

Reworked the API port test to fix some issues it was having.

Fixed a typo in the API version of the playerinfo function that caused the code to break if the bot was set to check for level hacking.  Luckily that defaults to disabled so this probably effected nobody before I caught it.