Bug fixes and performance and timing changes
As part of continued efforts to reduce command spam from the bot, I have made several changes to when and how frequently commands are sent.

Added queue and timer to make the bot send chat commands to the web API at intervals rather than immediately.

The bot now sends llp commands when a player joins or leaves and once per day when the server is not busy.  The claim removal code has also had a major overhaul and only removes claims within 100 metres of players and checks for them within the bot's stored data every 60 seconds.  Now claims are only discovered when a player leaves after which they get removed when players are nearby.