Bug fixes, new command queue, misc changes
Added new command queue called persistentQueue.  Unlike the other queues it is not deleted when the bot is restarted.  Commands in it are only executed when the bot is connected and survive in the queue until the bot runs them.

Several commands have been updated to use the persistent queue.

Fixed an issue with one of the /location lobby commands always executing even if you actually wanted the other command.

When the server is empty, the bot periodically sends runs the mem command.  This is used to monitor the tick count and also keep tabs on the number of players online.  The bot was running it every minute but it can run a lot less frequently so I have moved it to the 10 minute timer.