Bug fixes mostly
Change to /disable watch alerts.  It now stops the bot PM'ing ingame admins about base raids and base protection bounces.

New command /set reserved slot time {minutes}
There is a completely rewritten slots system almost ready to release.  This command is for that so it won't do anything currently.

Fixed bug in /setbase command.  Used wrong variable steam instead of chatvars.playerid which prevented the command reporting when you couldn't set the base there. Also added a bit more info to explain why.

Fixed regression bug in the bot's dig commands.  There was an out by one error that was fixed then unfixed >.<  It is fixed again (OR IS IT?) xD

Added check for players being below a height of zero (except for staff).  Hopefully this will rescue players that fall through the world and don't get automatically rescued by the server.  This is usually due to a map glitch so this fix might not work anyway.  It is triggered every minute to stop it going nuts and spamming tele commands.

Added a small hack (:O  Smegs hacking!) to try to fix a server issue where a corrupt player profile can freeze and crash the server.  The bot sends the sa command when non-brand new players join.  I further limited it so it will do this once per player per real day so that it doesn't create tons of lag every time lots of players join all at once.  That might happen occasionally but should be rare and brief if it ever happens.

Also put several commands on temporary timers when a player joins so that the bot doesn't immediately spam several commands.  Instead they are paced out over 40 seconds.

Fixed (I hope) a bug in the welcome message code when a player joins the server.  In some circumstances the bot was skipping the welcome messages.

Restored a delay that was preventing players being sent to the lobby in a timely fashion.  It wasn't broken but the delay was unacceptably long.  It is much more fashionable now.