Bug fixes
Removed the command /enable/disable claim scan.  It is no longer relevant or used since the bot scans for claims much more efficiently so there's no noticeable performance difference having the bot read claims or not read them.

The bug that is causing the bot to forget what mods are installed is still occurring though its not happening everywhere and not quickly.  I have made some changes to better detect and fix that and other changes that may stop it occuring in the first place.  It is a weird bug that shouldn't happen in the first place.

The bug only effects bots using the new API mode and as yet another way to catch and fix it failing to read the API, the bot checks for empty data every minute as well as when attempting to read the API data.  If it detects that the data is missing, it resends the effected command.  The resend is on a 60 second timer to stop it spamming too quickly.