Bug fixes
There seems to be a bug in the player archiver so until I get to it I have disabled the /archive players command as it isn't supposed to archive everyone :O

There are some almost finished changes to the bot's backup system.  The backups aren't changed but you'll be able to list them and restore older backups and also just restore bases and cash or just cash.  For the moment I am holding back some of the changes.

Fixed a bug in the bot's undo of spawned prefabs and also fixed an out by 1 issue with the vertical placement of prefabs.

The bot's daily inventory log now includes the game day and time.

The /lobby {player} command which is admin only will now send a player to the lobby regardless of the actual name of the location that is the lobby.

Players typing /lobby will be sent to whichever location is the lobby regardless of its true name.

Fixed a command clash where /location lobby {location name} was triggering if you actually typed /location lobby disable/enable waypoints.  It was only a problem if the location was called lobby which it usually is.

Players could teleport to their waypoint if it was in a location that disabled waypoints.  Now it correctly blocks the teleport.  It already blocked setting waypoints but did nothing to existing ones in the location.

Fixed a bug that prevented rolling announcements working.

Moved processing of LKP to a timed queue so that the bot processes 1 line per second instead of all together.  On servers with very large numbers of players it was causing the bot to 'freeze' while it processed LKP.  It should no longer freeze.

Changed the order that the bot loads scripts and custom scripts and added extra reporting so that an error in a custom script doesn't completely break the bot.  It will report the error and commands such as /reload code will still work even if the custom code is a complete disaster.

Moved the persistent queue to a slower timer (that runs every second instead of every half second).  Also fixed a bug in it that caused the queue to break if there was a teleport command and the player to be teleported had left the server.  Now it will delete the command and process the rest of the queue.

I undid a previous change that was intended to hide telnet messages from Mudlet's terminal window.  It made it too difficult to tell if the bot was connected to the server or not.  It was hiding too much stuff and depriving us of feedback.

I also disabled a recent feature where the bot would send a save world command when players joined.  That is brilliant for fixing corrupt player profiles but it also caused too much lag.  It will return as something you can target on a specific player to fix their profile.

Fixed a cosmetic issue where the bot was reporting the map limit to donors using the limit that applies to everyone else.  Donors get to travel an extra 5km.