Bug fixes, improvements
Fixed an issue that was causing a lot of server lag.  It wasn't a bug in the bot but appears to be a bug in the server.  Won't know if fixed in A17 or not for a few more days, but it has made a big difference in A16.

Removed a lot of unnecessary maths now that player coords are converted to Ints in telnet mode and are already Ints in API mode. For the moment I have added that same maths to the teleport function in case there are any floats in old data.

I improved several aspects of the bot's dig and fill and prefab spawn commands.  You can now change dig direction up and down as well as other directions.  Also the undo command is no longer recorded as your last command which means you can undo a bad dig/placement and redo after moving a bit.

Reverted an earlier change where Mudlet was suppressing all output to its telnet console window.  Some people actually use that so its all back again.

Fixed a bug in timeout where admins could get stuck there (at least its a safe place away from players).

Added better feedback if the bot detects that Alloc's mod is missing or not working.  It is required for sending PM's to players and for several other important console commands not included in the server.  Also the bot will automatically switch from using Alloc's web API back to using telnet if the API isn't working.

I improved the bot's DNS lookup.  Now it records every IP range, not just the Chinese ones.  These are searched before the bot does any new DNS lookups.  It will only query DNS servers if it sees an IP that doesn't match any known IP range. Also the bot and any other linked bots are limited to a maximum of 500 DNS lookups online per day to avoid getting blacklisted for doing too many DNS lookups in a day.

Added cash to some irc commands.  I will update the command help soon.

Added new table, bases.  This will replace the base information currently stored in the player record and will allow for servers to set how many bases an individual or class of player (eg donors) can have.  Current there's a hardcoded limit of 2 bases.  I haven't added code for this yet but that is coming.

Moved the SA console command from the 2 minute timer to the 5 minute timer as part of efforts to improve server performance by not spamming commands faster than necessary.

Fixed a bug in the bot's welcome message code that was preventing it welcoming players.

Fixed a bug in the bot's teleport parser that was incorrectly flagging players fetched by admins via the console as being suspicious teleports and if they did too many of them, the player would get banned. The bug mainly occured when the player or admin had a multi-part name.  The tele command is correctly parsed now.