Bug fixes, improvements and tweaks for A17 readiness
The bot's Coppi (and Patron's mod) support has been improved with some new commands.

When admins use the /mark command and follow up with the /save command, the info gets stored in a table.  Admins can list the contents of it with /list saves.  The list is numbered and entries can be deleted with /delete save {number}.  The marked area can also be teleported to (after /list saves) by typing /tp #{number from list}.  Also admins can now access the saves done by any admin, not just their own so there is no need to re-mark an area if it has previously been done.  This is particularly useful for applying or removing trader protection, but also applies to anything you can do with marked areas.

The beginnings of the move from having base information recorded in the player record to having it in a dedicated table is underway.  I may even finish the code today.  When completed, server owners will be able to set how many bases an individual or any class of player (eg donors) can have.

Fixed an issue where new bots on pvp servers weren't correctly marking the 4 map sectors as pvp.  They defaulted to pve and now they default to nothing and the bot sets them up.

Fixed compatibility with CSMM Patrons Mod. Two console commands had been renamed.

The bot can now detect when the map has been reset and will remind admins that the bot should also be reset.  This isn't a new feature but it had a silly logic bug in it.