API fixes, bug fixes, new commands and better compatibility with A16 and A17
Bug fixes:

Fixed typo in /hordeme command.

Fixed bugs in some of the bot's use of bc-block commands so that expected results actually occur.

Corrected use of exiled flag in player record which then required a hasty live fix as it resulted in the exiling of some players (mostly admins).  On reflection I should have left them there but I was worried about pitchforks. xD

Fixed wrong use of case in a call to a function in the one minute timer.  Lua is so case sensitive! :O

Misc changes and improvements:

Improved the bot's handling of detecting API online and telnet online and better handling of switching between them as necessary.

The bot no longer needs telnet and while it needs help getting connected to the API to start, it doesn't need to connect to telnet for anything.  There are advantages to keeping telnet enabled as a fallback or to help fix occasional issues that for some reason the bot can't fix using only the API (usually due to an un-fixed bug).

Block names used in dig commmands are now checked against the server and the bot automatically corrects the block name to match the valid block name known to the server (which uses mixed case).  Prefabs with few exceptions are all lowercase but blocknames must have correct case to spawn.

Updated the command help which is now included with the published bot scripts in the main scripts folder.

Changed the bot's use of bc-unlock all to target the player rather than the chunk as the BC command expects to come from a player.

Added a function that enables/disables several Mudlet triggers depending on if the bot is to use telnet or the API as the API handles calls to those triggers since telnet is not guaranteed to be enabled. More triggers may be toggled later as the bot running 100% with the API is still quite new.

The IRC command ( set player {name} cash {value} ) now accepts the word everyone instead of a player name.  This allows you to arbitrarily set every player's cash at once.

Added IRC command ( set update branch {branch name} ).  There is already an in-game version but an IRC version is useful for when in-game commands aren't working.  To use it, the bot must recognise you as a server owner.

Added IRC command ( enable/disable debug ).  This is mainly to help me debug the bot.  It turns on/off debug output in Mudlet's windows.  Self-hosters may find this useful.

Added IRC command ( set bot owner {steam ID} ).  This will be used later on.  At the moment setting it has no purpose.

Changed the API log polling interval default to 3.  New bots using the API will read new server log lines every 3 seconds.  You can set this to whatever you like but it directly effects how quickly the bot reacts to server messages and game chat.  The most I would consider setting it to is 5.