Bug fixes, new commands
Fixed bug that caused the bot to not read the gametime in API mode.

The shop can now be told how many units of an item to give to a player when they buy 1x the item.  Also the shop now applies the correct item quality if one is set rather than always being random quality.  Only items with no quality set will be random.

New IRC command: shop units {item} {new unit quantity}
Change the number of units sold per sale of the specified item.

Changed IRC command shop add item {item name} category {category} price {price} stock {max stock} units {units dropped}
Add an item to the shop.  If a unit is given and a player buys 1 item, the bot will give 1 * the unit eg. 10 of the item.

Shop item units can bet set/changed in-game too.

Players will see the units when they view the shop items before purchasing them.

Removed the delay before the bot teleports a player to the lobby so that they will be teleported as soon as the bot sees that they have spawned into the world.  This should reduce the number of instances where a new player places their starter kit on the ground then gets kidnapped.

Fixed a bug causing the bot to not parse PVP kills in A17.

In API mode the bot now saves a new log which is effectively the old telnet log but as seen via the API.