Bug fixes, performance adjustments and interface code for the new panel.
Changed default API log read interval from 1 second to 3 seconds.

The bot's automatic lag check timing is now configurable with the command /set lag check {seconds}.  The default is 15 seconds.
Additionally the bot will only suspend bot functions for a maximum of 10 seconds.  This is to prevent excessive lag whether real or not from causing the bot to stop accepting commands full time.

Added location info to the daily inventory log.

Fixed a bug that caused the bot to not accept ingame commands from IRC unless the bot was connected to the server.  This allows you to send what are normally ingame commands to the bot to change settings etc while the bot is offline.

Added command /set telnet enabled/disabled.  This just tells the bot if the server has telnet enabled or not so it won't keep endlessly trying to connect if there is no telnet to connect to.

Made some improvements to how the bot monitors the API and telnet so that it won't flag telnet as offline if it is infact still working when the bot is in API mode.  There is currently an issue with the API skipping some server log lines.  It appears to be a fault in the API as far as I can tell.  If your bot regularly needs to have commands repeated, you may want to switch it to using telnet.  The command for that is use telnet.

The /pack and /revive command now sends the player to the surface.  This is to fix some reported issues where the bot would sometimes teleport players above the ground.

In API mode, almost all Mudlet triggers are disabled.  Switching to telnet mode re-enables them.

Fixed a typo in the bot's /restore backup command that caused it to not restore anything.

Added code to enable the bot to communicate with the new web panel and for the panel to communicate with the bot.  The panel is able to access the bot's in-memory Lua tables and the bot's database and can do stuff like make live changes to stuff, restart the bot and server and command the bot to do anything that the bot has existing code for.  It can directly edit any table and tell the bot to reload the table.

Fixed typo in player_info.lua

Added new trigger, telnetCheck to flag the telnet status and timestamp of the last communication with the server's telnet for monitoring purposes.